requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
flask-restplus==0.13.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
pymongo==3.9.0
pyrsistent==0.15.5
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.6
Werkzeug==0.16.0
zipp==0.6.0
